extract classy events tests
