+++ tests/Events/ObjectTypeEventsTest.php
@@ -1,0 +1,108 @@
+<?php
+
+namespace Illuminate\Tests\Events;
+
+use Illuminate\Events\Dispatcher;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+
+class ObjectTypeEventsTest extends TestCase
+{
+    protected function tearDown(): void
+    {
+        m::close();
+    }
+
+    public function testClassesWork()
+    {
+        unset($_SERVER['__event.test']);
+        $d = new Dispatcher;
+        $d->listen(ExampleEvent::class, function () {
+            $_SERVER['__event.test'] = 'baz';
+        });
+        $d->dispatch(new ExampleEvent);
+
+        $this->assertSame('baz', $_SERVER['__event.test']);
+    }
+
+    public function testArrayCallbackListenersAreHandled()
+    {
+        unset($_SERVER['__event.ExampleListener']);
+        $d = new Dispatcher;
+        $d->listen(ExampleEvent::class, [ExampleListener::class, 'hear']);
+        $d->dispatch(new ExampleEvent);
+
+        $this->assertTrue($_SERVER['__event.ExampleListener']);
+    }
+
+    public function testEventClassesArePayload()
+    {
+        unset($_SERVER['__event.test']);
+        $d = new Dispatcher;
+        $d->listen(ExampleEvent::class, function ($payload) {
+            $_SERVER['__event.test'] = $payload;
+        });
+        $d->dispatch($e = new ExampleEvent, ['foo']);
+
+        $this->assertSame($e, $_SERVER['__event.test']);
+    }
+
+    public function testInterfacesWork()
+    {
+        unset($_SERVER['__event.test']);
+        $d = new Dispatcher;
+        $d->listen(SomeEventInterface::class, function () {
+            $_SERVER['__event.test'] = 'bar';
+        });
+        $d->dispatch(new AnotherEvent);
+
+        $this->assertSame('bar', $_SERVER['__event.test']);
+    }
+
+    public function testBothClassesAndInterfacesWork()
+    {
+        unset($_SERVER['__event.test']);
+        $_SERVER['__event.test'] = [];
+        $d = new Dispatcher;
+        $d->listen(AnotherEvent::class, function ($p) {
+            $_SERVER['__event.test'][] = $p;
+            $_SERVER['__event.test1'] = 'fooo';
+        });
+        $d->listen(SomeEventInterface::class, function ($p) {
+            $_SERVER['__event.test'][] = $p;
+            $_SERVER['__event.test2'] = 'baar';
+        });
+        $d->dispatch($e = new AnotherEvent, ['foo']);
+
+        $this->assertSame($e, $_SERVER['__event.test'][0]);
+        $this->assertSame($e, $_SERVER['__event.test'][1]);
+        $this->assertSame('fooo', $_SERVER['__event.test1']);
+        $this->assertSame('baar', $_SERVER['__event.test2']);
+
+        unset($_SERVER['__event.test1']);
+        unset($_SERVER['__event.test2']);
+    }
+}
+
+class ExampleEvent
+{
+    //
+}
+
+interface SomeEventInterface
+{
+    //
+}
+
+class AnotherEvent implements SomeEventInterface
+{
+    //
+}
+
+class ExampleListener
+{
+    public function hear()
+    {
+        $_SERVER['__event.ExampleListener'] = true;
+    }
+}
